Refactor platform-specific functionality.
Things are much more extensible this way.
